Backend_FastAPI/requirements.txt
fastapi
uvicorn[standard]
requests
python-docx
pymupdf
passlib
bcrypt<4.1
python-dotenv
openai
chromadb
langchain
langchain-openai
tiktoken
pillow
numpy
pytesseract
pdf2image

# Document Indexing Service Dependencies
watchdog>=4.0.0
pdfminer.six>=20231228
pdf2image>=1.17.0
pytesseract>=0.3.10
camelot-py[cv]>=0.11.0
rank-bm25>=0.2.2
tiktoken>=0.5.0

# System packages required on host:
# sudo apt-get install -y poppler-utils tesseract-ocr ghostscript
